NamoNexus/requirements.txt
python-dotenv
termcolor
psutil
sentence-transformers
faiss-cpu
numpy
